remove old tests and old flask code
