+++ jskit/fasttext/config.py
