+++ jskit/fasttext/example_training_data.json
