+++ jskit/fasttext/json_to_train.py
