--- jskit/fasttext/tests/test_sentences.txt
@@ -1,5 +1,0 @@
-what can you do
-hi greetings to you
-i'll see you later
-which documents are necessary for entry
-blah blah zorg porp
